Funktioniert!

+++ Program.cs
@@ -2,6 +2,6 @@
 Console.WriteLine(":. Mein erstes Projekt auf GitHub .:");
 int i = 1;
 int j = 2;
-int k;
-k = i / j;
+double k;
+k = (double)i / j;
 Console.WriteLine($"Wert von k: {k}");
